[Feature #163002540] Refactor code to meet specification
