+++ app/api/v1/admin/routes.py
@@ -1,7 +1,7 @@
 """The meetup routes"""
 
 from flask import jsonify, request , make_response , abort
-from app.admin.models import MeetupModel, MEETUPS_LEN, QuestionModel , CommentModel, COMMENTS_LEN
+from app.admin.models import MeetupModel, MEETUPS_LEN
 from app.api.v1 import path_1 
 
 @path_1.route("/meetups", methods=['POST'])
@@ -84,19 +84,3 @@ def meetup_rsvp(meetup_id, resp):
                                                'topic':meetup['topic'],
                                                'Attending':resp}]}), 200
 
-#user post comment
-@path_1.route("/questions/<int:question_id>/comment", methods=['POST'])
-def user_comment_on_a_question(question_id):
-    """
-    User post comment endpoint route
-    """
-    try:
-        comment = request.get_json()['comment']
-    except KeyError:
-        abort(make_response(jsonify({'status': 400, 'error':'Check your json key. It is comment'})))
-    question = QuestionModel.get_question(question_id)
-    if question:
-        my_question = question[0]
-        my_question['comments'].append(comment)
-        return jsonify({"status": 201, "data": my_question}), 201
-    return jsonify({'status': 404, 'error':'The question you are looking for is not found'}), 404
